Change project structure
- Create Duke package
- Create UI package
- Move CLI UI code to UI package
