--- src/main/java/Duke.java
@@ -1,14 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-        String greeting = "------------------------------------------------------------\n" +
-                "Hello! I'm Ciara\n" +
-                "What can I do for you?\n";
-
-        String goodbye = "------------------------------------------------------------\n" +
-                "Bye. Hope to see you again soon!\n" +
-                "------------------------------------------------------------";
-
-        System.out.println(greeting);
-        System.out.println(goodbye);
-    }
-}
