+++ src/main/java/duke/Duke.java
@@ -1,0 +1,9 @@
+package duke;
+
+import duke.ui.Cli;
+
+public class Duke {
+    public static void main(String[] args) {
+        Cli.printUI();
+    }
+}
